refactor(crawler): modularize crawl controls
